PlanbookAI/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pika==1.3.2
redis==5.0.1
PyJWT==2.8.0
Pillow==10.1.0
boto3==1.34.0
python-multipart==0.0.6




